Another workaround for a truly bizzare bug
I have yet to figure out why, but it appears that every once in a while the random number generator spits out a garbage value for a few dozen calls into this specific method. I'm detecting this and falling back to a simple not-random-but-the-player-will-never-know solution pulling values from a predefined random number table. Will hopefully figure out the root cause and fix that eventually, but until then this stops the game from crashing and produces reasonable-looking results.